Multiple updates
Better handling of multiple clients
Do not override key pairs if they are already on the server
Better handling of alternate config file locations

+++ server-setup.sh
@@ -1,9 +1,11 @@
 # Some variables you may want to change
 PRIVATE_ADDRESS_OF_SERVER=10.20.10.1
-PRIVATE_ADDRESS_OF_LOCAL_CLIENT=10.20.10.2
-GAME_SERVER_PORTS=2456,2459
+# Example: client_ip1,port,port-client_ip2,port,port-client_ip3-client_ip4,port
+# Forwards the port on the server to the same port on the client.
+PRIVATE_ADDRESS_AND_PORTS_OF_LOCAL_CLIENTS=10.20.10.2,2456,2459-10.20.10.3
 WIREGUARD_PORT=51820
 WIREGUARD_CONFIG_FILE=/etc/wireguard/wg0.conf
+ALLOWED_CLIENT_IPS=10.20.10.2/32,10.20.10.3/32
 
 # Install Wireguard and Generate Keys
 sudo apt update
@@ -12,14 +14,18 @@ sudo apt install vim
 sudo apt install resolvconf
 
 # Server keys
-wg genkey | sudo tee /etc/wireguard/server-private.key
-sudo chmod go= /etc/wireguard/server-private.key
-sudo cat /etc/wireguard/server-private.key | wg pubkey | sudo tee /etc/wireguard/server-public.key
+if sudo test ! -e "/etc/wireguard/server-private.key"; then
+  wg genkey | sudo tee /etc/wireguard/server-private.key
+  sudo chmod go= /etc/wireguard/server-private.key
+  sudo cat /etc/wireguard/server-private.key | wg pubkey | sudo tee /etc/wireguard/server-public.key
+fi
 
 # Client keys
-wg genkey | sudo tee /etc/wireguard/client-private.key
-sudo chmod go= /etc/wireguard/client-private.key
-sudo cat /etc/wireguard/client-private.key | wg pubkey | sudo tee /etc/wireguard/client-public.key
+if sudo test ! -e "/etc/wireguard/client-private.key"; then
+  wg genkey | sudo tee /etc/wireguard/client-private.key
+  sudo chmod go= /etc/wireguard/client-private.key
+  sudo cat /etc/wireguard/client-private.key | wg pubkey | sudo tee /etc/wireguard/client-public.key
+fi
 
 # Store the things we need to build the config files
 SERVER_PRIVATE_KEY=$(sudo cat /etc/wireguard/server-private.key)
@@ -39,19 +45,31 @@ echo "ListenPort = $WIREGUARD_PORT" | sudo tee -a $WIREGUARD_CONFIG_FILE
 echo "" | sudo tee -a $WIREGUARD_CONFIG_FILE
 echo "PostUp = ufw route allow in on wg0 out on $INTERNET_INTERFACE" | sudo tee -a $WIREGUARD_CONFIG_FILE
 echo "PostUp = iptables -t nat -I POSTROUTING -o $INTERNET_INTERFACE -j MASQUERADE" | sudo tee -a $WIREGUARD_CONFIG_FILE
-if [ "$GAME_SERVER_PORTS" ]; then
-  echo "PostUp = iptables -t nat -A PREROUTING -i $INTERNET_INTERFACE -p tcp -m multiport --dports $GAME_SERVER_PORTS -j DNAT --to-destination $PRIVATE_ADDRESS_OF_LOCAL_CLIENT" | sudo tee -a $WIREGUARD_CONFIG_FILE
-  echo "PostUp = iptables -t nat -A PREROUTING -i $INTERNET_INTERFACE -p udp -m multiport --dports $GAME_SERVER_PORTS -j DNAT --to-destination $PRIVATE_ADDRESS_OF_LOCAL_CLIENT" | sudo tee -a $WIREGUARD_CONFIG_FILE
-fi
+
+# For each client, forward any ports specified in PRIVATE_ADDRESS_AND_PORTS_OF_LOCAL_CLIENTS.
+CLIENTS_WITH_PORTS="$(echo "$PRIVATE_ADDRESS_AND_PORTS_OF_LOCAL_CLIENTS" | cut -d "-" -f 1- --output-delimiter=' ')"
+for client_with_ports in $(echo $CLIENTS_WITH_PORTS); do
+  CLIENT="$(echo "$client_with_ports" | cut -d "," -f 1)"
+  PORTS="$(echo "$client_with_ports" | cut -d "," -f 2- --output-delimiter=',')" # Comma delimiter for dports parameter
+  if [ "$PORTS" ] && [ "$CLIENT" ] && [ "$PORTS" != "$CLIENT" ]; then
+    echo "PostUp = iptables -t nat -A PREROUTING -i $INTERNET_INTERFACE -p tcp -m multiport --dports $PORTS -j DNAT --to-destination $CLIENT" | sudo tee -a $WIREGUARD_CONFIG_FILE
+    echo "PostUp = iptables -t nat -A PREROUTING -i $INTERNET_INTERFACE -p udp -m multiport --dports $PORTS -j DNAT --to-destination $CLIENT" | sudo tee -a $WIREGUARD_CONFIG_FILE
+  fi
+done
+
 echo "PreDown = ufw route delete allow in on wg0 out on $INTERNET_INTERFACE" | sudo tee -a $WIREGUARD_CONFIG_FILE
 echo "PreDown = iptables -t nat -D POSTROUTING -o $INTERNET_INTERFACE -j MASQUERADE" | sudo tee -a $WIREGUARD_CONFIG_FILE
-if [ "$GAME_SERVER_PORTS" ]; then
-  echo "PreDown = iptables -t nat -D PREROUTING -i $INTERNET_INTERFACE -p tcp -m multiport --dports $GAME_SERVER_PORTS -j DNAT --to-destination $PRIVATE_ADDRESS_OF_LOCAL_CLIENT" | sudo tee -a $WIREGUARD_CONFIG_FILE
-  echo "PreDown = iptables -t nat -D PREROUTING -i $INTERNET_INTERFACE -p udp -m multiport --dports $GAME_SERVER_PORTS -j DNAT --to-destination $PRIVATE_ADDRESS_OF_LOCAL_CLIENT" | sudo tee -a $WIREGUARD_CONFIG_FILE
-fi
+for client_with_ports in $(echo $CLIENTS_WITH_PORTS); do
+  CLIENT="$(echo "$client_with_ports" | cut -d "," -f 1 --output-delimiter=' ')"
+  PORTS="$(echo "$client_with_ports" | cut -d "," -f 2- --output-delimiter=',')" #Comma delimiter for dports parameter
+  if [ "$PORTS" ] && [ "$CLIENT" ] && [ "$PORTS" != "$CLIENT" ]; then
+    echo "PreDown = iptables -t nat -D PREROUTING -i $INTERNET_INTERFACE -p tcp -m multiport --dports $PORTS -j DNAT --to-destination $CLIENT" | sudo tee -a $WIREGUARD_CONFIG_FILE
+    echo "PreDown = iptables -t nat -D PREROUTING -i $INTERNET_INTERFACE -p udp -m multiport --dports $PORTS -j DNAT --to-destination $CLIENT" | sudo tee -a $WIREGUARD_CONFIG_FILE
+  fi
+done
 echo "" | sudo tee -a $WIREGUARD_CONFIG_FILE
 echo "[Peer]" | sudo tee -a $WIREGUARD_CONFIG_FILE
-echo "AllowedIPs = $PRIVATE_ADDRESS_OF_LOCAL_CLIENT/32" | sudo tee -a $WIREGUARD_CONFIG_FILE
+echo "AllowedIPs = $ALLOWED_CLIENT_IPS" | sudo tee -a $WIREGUARD_CONFIG_FILE
 echo "# local client public key" | sudo tee -a $WIREGUARD_CONFIG_FILE
 echo "PublicKey = $CLIENT_PUBLIC_KEY" | sudo tee -a $WIREGUARD_CONFIG_FILE
 
@@ -67,14 +85,18 @@ sudo ufw allow 53/udp
 sudo ufw allow 22/tcp
 sudo ufw allow OpenSSH
 
-if [ "$GAME_SERVER_PORTS" ]; then
-  for i in $(echo "$GAME_SERVER_PORTS" | cut -d "," -f 1- --output-delimiter=' '); do
-    sudo ufw allow $i/udp
-    sudo ufw allow $i/tcp
-    sudo ufw route allow proto udp to $PRIVATE_ADDRESS_OF_LOCAL_CLIENT port $i
-    sudo ufw route allow proto tcp to $PRIVATE_ADDRESS_OF_LOCAL_CLIENT port $i
-  done
-fi
+for client_with_ports in $(echo $CLIENTS_WITH_PORTS); do
+  CLIENT="$(echo "$client_with_ports" | cut -d "," -f 1)"
+  PORTS="$(echo "$client_with_ports" | cut -d "," -f 2- --output-delimiter=' ')" # Comma delimiter for dports parameter
+  if [ "$PORTS" ] && [ "$CLIENT" ] && [ "$PORTS" != "$CLIENT" ]; then
+    for port in $(echo $PORTS); do
+      sudo ufw allow $port/udp
+      sudo ufw allow $port/tcp
+      sudo ufw route allow proto udp to "$CLIENT" port "$port"
+      sudo ufw route allow proto tcp to "$CLIENT" port "$port"
+    done
+  fi
+done
 
 # If ufw is on, then this disable never seems to happen instanly, so I put a sleep in here. Shoot me.
 sudo ufw disable
